WIP: alternative implementation of wasm store without wasmtime
Given the quite large binary size of wasmtime (5.8MB after purposefully
minifying it) we wanted to investigate the use of alternative wasm
runtimes.

I tried to sketch this out using the wasm-c-api interface. Unfortunately
this didn't work wery well. wasm-micro-runtime looked as a very good fit
for us, and ostensibly uses the wasm.h api but some functionality is
missing, like host creation of tables. Wasmtime itself appears to have
quite a few bugs in the wasm-c-api, likely because everyone uses its own
C API. Wasm3 which is a portable interpreter, also has its own API.

A more pragmatic approach is to allow alternative implementations
of wasm_store.c , which looked quite feasible as this is already quite
isolated from implementation details of other parts of the tree-sitter
runtime library (except for needing to reconstruct a TSLanguage struct)

I sketched this out using my own toy wasm interpreter, which isn't
robust or fast enough yet, but made it easier to debug and understand
all whats going on.

As this is partially written in zig, it can be tried out using the
https://github.com/neovim/neovim/pull/28344 branch with the -Dwasm flag.

Some changes to the design to make re-implementation easier

- Each language has its own private wasm memory and functable (there is
  really no shared state between languages at all. we just need a stub
  TSWasmStore to match the public API).
- implement stdlib functions on the host side instead of shipping a compiled wasm stdlib
- serialize/deserialize can be implemented by saving/restoring the heap
  (which now only contains one scanner object)
- simplify copying of static language data. instead of allocating a lot
  of individual TSLanguage arrays and strings, copy the entire static
  data block at once. Then use pointers into this block